requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.7.0
django-heroku==0.3.1
django-toolbelt==0.0.1
djangorestframework==3.12.2
facebook-sdk==3.1.0
gunicorn==20.0.4
idna==2.10
instagram-basic-display==1.0.5
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
wcwidth==0.2.5
whitenoise==5.2.0
redis==3.5.3
Celery==5.0.5
